+++ p5-analysis/src/index.ts
@@ -1,4 +1,5 @@
 export { Category } from './models/Category';
+export { Cdn } from './models/Cdn';
 export { Library } from './models/Library';
 export { Script } from './models/Script';
 export { Sketch, SketchStructureType } from './models/Sketch';